tests: move _databases_support_transactions implementation to mixin
This avoids copying this code to multiple tests.
