+++ weblate/trans/tests/test_autotranslate.py
@@ -11,7 +11,7 @@ from django.urls import reverse
 
 from weblate.trans.models import Component
 from weblate.trans.tests.test_views import ViewTestCase
-from weblate.utils.db import using_postgresql
+from weblate.utils.db import TransactionsTestMixin
 
 
 class AutoTranslationTest(ViewTestCase):
@@ -166,17 +166,7 @@ class AutoTranslationTest(ViewTestCase):
             call_command("auto_translate", "test", "test", "xxx")
 
 
-class AutoTranslationMtTest(ViewTestCase):
-    @classmethod
-    def _databases_support_transactions(cls):
-        # This is workaround for MySQL as FULL TEXT index does not work
-        # well inside a transaction, so we avoid using transactions for
-        # tests. Otherwise we end up with no matches for the query.
-        # See https://dev.mysql.com/doc/refman/5.6/en/innodb-fulltext-index.html
-        if not using_postgresql():
-            return False
-        return super()._databases_support_transactions()
-
+class AutoTranslationMtTest(TransactionsTestMixin, ViewTestCase):
     def setUp(self):
         super().setUp()
         # Need extra power
