+++ weblate/trans/tests/test_selenium.py
@@ -37,7 +37,7 @@ from weblate.trans.tests.utils import (
     get_test_file,
     social_core_override_settings,
 )
-from weblate.utils.db import using_postgresql
+from weblate.utils.db import TransactionsTestMixin
 from weblate.vcs.ssh import get_key_data
 from weblate.wladmin.models import ConfigurationError, SupportStatus
 
@@ -63,21 +63,13 @@ SOURCE_FONT = os.path.join(
 )
 
 
-class SeleniumTests(BaseLiveServerTestCase, RegistrationTestMixin, TempDirMixin):
+class SeleniumTests(
+    TransactionsTestMixin, BaseLiveServerTestCase, RegistrationTestMixin, TempDirMixin
+):
     driver = None
     driver_error = ""
     image_path = None
     site_domain = ""
-
-    @classmethod
-    def _databases_support_transactions(cls):
-        # This is workaround for MySQL as FULL TEXT index does not work
-        # well inside a transaction, so we avoid using transactions for
-        # tests. Otherwise we end up with no matches for the query.
-        # See https://dev.mysql.com/doc/refman/5.6/en/innodb-fulltext-index.html
-        if not using_postgresql():
-            return False
-        return super()._databases_support_transactions()
 
     @contextmanager
     def wait_for_page_load(self, timeout=30):
