+++ weblate/utils/db.py
@@ -28,6 +28,18 @@ def using_postgresql():
     return connection.vendor == "postgresql"
 
 
+class TransactionsTestMixin:
+    @classmethod
+    def _databases_support_transactions(cls):
+        # This is workaround for MySQL as FULL TEXT index does not work
+        # well inside a transaction, so we avoid using transactions for
+        # tests. Otherwise we end up with no matches for the query.
+        # See https://dev.mysql.com/doc/refman/5.6/en/innodb-fulltext-index.html
+        if not using_postgresql():
+            return False
+        return super()._databases_support_transactions()
+
+
 def adjust_similarity_threshold(value: float):
     """
     Adjusts pg_trgm.similarity_threshold for the % operator.
